Improve zabbix performance using get asynchronous mode
Zabbix version 7.0.0 introduces asynchronous get as a
replacement for the traditional synchronous get method.

Using the traditional synchronous get mode, the poller
waits for a response from the device to fulfill a new
request. Even when increasing the number of pollers
in scenarios with many hosts and items, the poller
process will experience high resource consumption.

Signed-off-by: assmann <assmann@datacom.com.br>